Assistant checkpoint: Add package required categories aggregation
Assistant generated file changes:
- client/src/hooks/use-services.ts: Add required categories aggregation for packages

---

User prompt:

We have the ability to create packages in our services page where we can combine multiple services to create a package which can be used as a service when booking an appointment. The issue is that the complete service form is unable to show the required categories for the the pack. all other fields are working correctly and I was able to save inventory by using the Add Item button under additional items used. Do we need to fix this or is it too much of an undertaking?