Assistant checkpoint: Update time slot availability check to support multiple groomers
Assistant generated file changes:
- client/src/hooks/use-appointments.ts: Fix time slot availability check to consider groomers

---

User prompt:

That worked, i was not able to book an appointment for 11 am as it was already booked and was then able to create an appointment for 11:30 am which was auto assigned to groomer one. When i try to book another appointment for the same time i.e. 11:30 am it does not allow me, however since we have 2 groomers it should allow me to book the appointment for 11:30 am and auto assign to groomer two.