Assistant checkpoint: Updated ServiceHistory to use correct usedItems property
Assistant generated file changes:
- client/src/components/ServiceHistory.tsx: Fix used items data fetching

---

User prompt:

you are not using the past chat context which is causing this confusion. the productsUsed null needs to be ignore. We are storing usedItems in service_history collection. The inventory collection has the name of the itemId that needs to be shown. Its very simple, no?